FastAPI_app_1/requirements.txt
fastapi
uvicorn
diffusers
torch
transformers
Pillow
python-multipart